+++ Src/Main/Modules/Memory_Managers/memory_free.c
@@ -191,7 +191,7 @@ void free_matrix(Matrix_T *m)
     
   if (m->reg_f)
   {
-    free_2d_mem(m->reg->A,m->row);
+    free_2d_mem(m->reg->A,(long unsigned)m->row);
   }
   else if (m->tri_f)
   {
@@ -249,5 +249,6 @@ void free_matrix(Matrix_T *m)
   }
   else
     abortEr("No matrix format is defined for this given matrix.\n");
+    
   free(m);
 }
